Added trailing slashes to permalinks
Clean URL:s in Jekyll are folder based and paths to folders should always have a trailing slash since otherwise the path is pointing to a file. For some reason I've missed adding that trailing slash in my permalink structure, but thanks to @ptz0n bringing the issue to my attention I have now fixed it.

Also - to make sure that any error on my behalf regarding the paths shouldn't have to result in needless duplicates in consumers of my page I've now added rel-canonical links that are always pointing to the main URL of the page.